feat: create edit and delete methods into admin controller
