+++ laracasts/laravel/laravel-project/app/Http/Controllers/AdminController.php
@@ -3,10 +3,15 @@
 namespace App\Http\Controllers;
 
 use App\Models\Post;
-use Illuminate\Http\Request;
+use Illuminate\Contracts\View\Factory;
+use Illuminate\Contracts\View\View;
+use Illuminate\Validation\Rule;
 
 class AdminController extends Controller
 {
+    /**
+     *  Show all posts
+     */
     public function index()
     {
         return view('admin.posts.index', [
@@ -14,13 +19,69 @@ class AdminController extends Controller
         ]);
     }
 
-    public function update()
-    {
-        return;
+    /**
+     * Return the form to create a post
+     */
+    public function create(): View|Factory
+    {    
+        return view('admin.posts.create');
     }
 
-    public function delete()
+    /**
+     * Create a post into database
+     */
+    public function store()
     {
-        return;
+        $attributes = request()->validate([
+            'title' => 'required',
+            'thumbnail' => 'image',
+            'slug' => ['required', Rule::unique('posts', 'slug')],
+            'excerpt' => 'required',
+            'body' => 'required',
+            'category_id' => ['required', Rule::exists('categories', 'id')]
+        ]);
+
+        $attributes['user_id'] = auth()->id();
+        $attributes['thumbnail'] = request()->file('thumbnail') ? request()->file('thumbnail')->store() : null;
+
+        Post::create($attributes);
+
+        return redirect('/')
+            ->with('success', 'Your post has been submmited');
+    }
+
+    public function edit(Post $post)
+    {
+        return view('admin.posts.edit',[
+            'post' => $post
+        ]);
+    }
+
+    public function update(Post $post)
+    {
+        $attributes = request()->validate([
+            'title' => 'required',
+            'thumbnail' => 'image',
+            'slug' => ['required', Rule::unique('posts', 'slug')->ignore($post->id)],
+            'excerpt' => 'required',
+            'body' => 'required',
+            'category_id' => ['required', Rule::exists('categories', 'id')]
+        ]);
+
+        if (isset($attributes['thumbnail'])) {
+            $attributes['thumbnail'] = request()->file('thumbnail') ? request()->file('thumbnail')->store() : null;
+        }
+
+        $post->update($attributes);
+
+        return back()->with('success', 'Post Updated !');
+    }
+
+    public function destroy(Post $post)
+    {
+        $post->delete();
+
+        return back()
+            ->with('success', 'Post Deleted');
     }
 }
