+++ laracasts/laravel/laravel-project/app/Http/Controllers/PostController.php
@@ -20,36 +20,11 @@ class PostController extends Controller
         ]);
     }
 
-    public function create(): View|Factory
-    {    
-        return view('admin.posts.create');
-    }
-
     public function show(Post $post): View|Factory
     {
         return view('posts.show',
         [
             'post'=> $post
         ]);
-    }
-
-    public function store()
-    {
-        $attributes = request()->validate([
-            'title' => 'required',
-            'thumbnail' => 'image',
-            'slug' => ['required', Rule::unique('posts', 'slug')],
-            'excerpt' => 'required',
-            'body' => 'required',
-            'category_id' => ['required', Rule::exists('categories', 'id')]
-        ]);
-
-        $attributes['user_id'] = auth()->id();
-        $attributes['thumbnail'] = request()->file('thumbnail') ? request()->file('thumbnail')->store() : null;
-
-        Post::create($attributes);
-
-        return redirect('/')
-            ->with('success', 'Your post has been submmited');
     }
 }
